Use AWS provided lambda image
It is more convenient to use the provided image
so that we get the Runtime Interface Emulator (RIE)
for free. This commit changes the base image and
download and install Julia 1.5.3 as part of the
Docker build.

A bootstrap shell script is added so that the
emulator can start the Julia program properly.

Also implements the Runtime Interface Client (RIC)
API fully - handling both invocation and initialization
errors.

Also enhance the sample package to do something
a little more meaningful (monte carlo pi calculation)
than hello-world.